Four conflicts, all resolved by keeping both sides' intent:

- The sidebar entry became a button on main and had been extracted into a
  helper here so pages can be attached at runtime. The helper now builds
  main's button, so an extension still appears without reopening the window
  and the entry is the widget main chose.
- A terminal's history extraction and its link handling moved into their own
  modules here while main edited them in place. The moved copies were
  identical apart from main's improvement to scheme detection, which is
  carried across: a person types WWW as readily as www, and refusing the
  capital would open nothing.
- Both sides taught the design lint to ignore a directory; it ignores both.
- Checkpoint imports were a straight superset.

Fixes three things the type check alone did not catch: a test that followed
its subject into another module, a bench window that predates main's zoom
state, and a manifest the extension built in code as well as shipping as a
file. The file is now the only description of that extension — two
descriptions had already drifted, the code one claiming an entrypoint, icon
and resources the shipped document does not.

587 tests pass across the workspace.

+++ src/runtime/hl-native/src/native/translator/guest/x86_64/rep_runtime.h
@@ -18,7 +18,7 @@ enum { X86_SOFT_READ = 1u, X86_SOFT_WRITE = 2u };
 
 struct cpu;
 
-typedef void (*hl_x86_rep_store_commit_fn)(uint64_t guest, uint64_t size);
+typedef void (*hl_x86_rep_store_commit_fn)(uint64_t guest, size_t size);
 typedef int (*hl_x86_rep_store_observation_active_fn)(void);
 typedef int (*hl_x86_rep_access_fn)(uint64_t guest, size_t size);
 typedef int (*hl_x86_rep_access_special_fn)(uint64_t guest, size_t size, int write);
